moves links to blank layout and simplifies
